requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.5
passlib==1.7.4
python-jose==3.5.0
redis==6.2.0
python-dotenv==1.1.0 